Starting the animation
Added a call to start() in init() to start the animation without calling it. And the docs don't say that we need to call it programatically doing mFunnyText.start(), so I conclude that the animation have to start automatically. If the animation has to start programatically please ignore this update and update the docs. Thanks